Mostly minor changes, changed the turns thing around a little bit to only increment on certain actions, also hopefully fixed a bug with the cave generation that was causing tiles in out of bounds areas to become floor.

+++ include/trl-main.h
@@ -126,7 +126,7 @@ typedef struct object_s {
 
 	uint8_t state;
 
-	uint8_t turns;
+	uint32_t turns;
 } object_t;
 
 typedef struct tile_s {
